moved blocks
