+++ lib/bloc/aspect/aspect_state.dart
