+++ lib/bloc/category/area/area_event.dart
