+++ lib/bloc/category/area/area_state.dart
