+++ lib/bloc/photo/photo_event.dart
