+++ lib/bloc/photo/photo_state.dart
