+++ lib/screens/aspect_distribution/aspects_to_distribute.dart
@@ -1,6 +1,6 @@
 import 'package:ehsfocus/models/aspect/aspects_model.dart';
 import 'package:ehsfocus/screens/aspect_distribution/audit_list.dart';
-import 'package:ehsfocus/screens/forms/aspects/bloc/aspect_bloc.dart';
+import 'package:ehsfocus/bloc/aspect/aspect_bloc.dart';
 import 'package:ehsfocus/services/popup_service/generic_message_popup.dart';
 
 import 'package:ehsfocus/shared/constants.dart';
