+++ lib/screens/aspect_distribution/aspects_to_fix_list.dart
@@ -1,5 +1,5 @@
 import 'package:ehsfocus/screens/aspect_distribution/audit_list.dart';
-import 'package:ehsfocus/screens/forms/aspects/bloc/aspect_bloc.dart';
+import 'package:ehsfocus/bloc/aspect/aspect_bloc.dart';
 import 'package:ehsfocus/shared/constants.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
