+++ lib/screens/aspect_distribution/rejected_aspects.dart
@@ -1,6 +1,6 @@
 import 'package:ehsfocus/models/aspect/aspects_model.dart';
 import 'package:ehsfocus/screens/aspect_distribution/audit_list.dart';
-import 'package:ehsfocus/screens/forms/aspects/bloc/aspect_bloc.dart';
+import 'package:ehsfocus/bloc/aspect/aspect_bloc.dart';
 import 'package:ehsfocus/shared/constants.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -9,7 +9,6 @@ class RejectedAspects extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     AspectBloc _aspect = AspectBloc();
-    var edit;
     return BlocProvider<AspectBloc>(
       create: (context) {
         _aspect.getMyRejectedAspects();
