+++ lib/screens/forms/area/area_providers.dart
@@ -1,7 +1,7 @@
 import 'package:ehsfocus/bloc/category/category_bloc.dart';
 import 'package:ehsfocus/screens/forms/area/area_service.dart';
-import 'package:ehsfocus/screens/forms/area/bloc/area_bloc.dart';
-import 'package:ehsfocus/screens/forms/employee/bloc/employee_bloc.dart';
+import 'package:ehsfocus/bloc/category/area/area_bloc.dart';
+import 'package:ehsfocus/bloc/employee/employee_bloc.dart';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
