+++ lib/screens/forms/aspects/photo_comment_form.dart
@@ -2,8 +2,8 @@ import 'package:ehsfocus/models/aspect/aspects_model.dart';
 import 'package:ehsfocus/bloc/category/category_bloc.dart';
 import 'package:ehsfocus/screens/category/category_picker.dart';
 import 'package:ehsfocus/screens/forms/area/are_dropdown_picker.dart';
-import 'package:ehsfocus/screens/forms/area/bloc/area_bloc.dart';
-import 'package:ehsfocus/screens/forms/audit/audit_bloc/audit_bloc.dart';
+import 'package:ehsfocus/bloc/category/area/area_bloc.dart';
+import 'package:ehsfocus/bloc/audit_bloc/audit_bloc.dart';
 import 'package:ehsfocus/shared/photoComponents/camera_service.dart';
 import 'package:ehsfocus/shared/action_button.dart';
 import 'package:ehsfocus/shared/form_eleements/audit_form_wraper.dart';
