+++ lib/screens/forms/employee/employee_service.dart
@@ -1,5 +1,5 @@
 import 'package:ehsfocus/models/generic_list_model.dart';
-import 'package:ehsfocus/screens/forms/employee/bloc/employee_bloc.dart';
+import 'package:ehsfocus/bloc/employee/employee_bloc.dart';
 
 class EmployeeService {
   List<GenericListObject> mapEmployeesToList(state) {
